model/requirements.txt
alembic==1.6.3
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.0.0
Flask-SQLAlchemy==2.5.1
fuzzywuzzy[speedup]
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
Mako==1.1.4
MarkupSafe==2.0.1
nltk==3.6.4
numpy==1.19.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
regex==2021.4.4
six==1.16.0
SQLAlchemy==1.4.15
tqdm==4.60.0
Werkzeug==2.0.1
PyMySQL==1.0.2
python-Levenshtein
transformers==4.18.0
textdistance
pandas
torch
typing
